♻️ Reorganisation modèle réponse signée
