+++ packages/applications/document-builder/src/modèleRéponseSignée/garantiesFinancières/modèleRéponseSignéeMiseEnDemeure.ts
@@ -1,0 +1,29 @@
+export const modèleRéponseMiseEnDemeureFileName =
+  'garanties-financières-modèle-mise-en-demeure.docx';
+
+export type ModèleMiseEnDemeure = {
+  type: 'mise-en-demeure';
+  data: {
+    dreal: string;
+    dateMiseEnDemeure: string;
+    contactDreal: string;
+    referenceProjet: string;
+    titreAppelOffre: string;
+    dateLancementAppelOffre: string;
+    nomProjet: string;
+    adresseCompleteProjet: string;
+    puissanceProjet: string;
+    unitePuissance: string;
+    titrePeriode: string;
+    dateNotification: string;
+    paragrapheGF: string;
+    garantieFinanciereEnMois: string;
+    dateFinGarantieFinanciere: string;
+    dateLimiteDepotGF: string;
+    nomRepresentantLegal: string;
+    adresseProjet: string;
+    codePostalProjet: string;
+    communeProjet: string;
+    emailProjet: string;
+  };
+};
