+++ packages/applications/document-builder/src/modèleRéponseSignée/modèleRéponseSignée.ts
@@ -1,0 +1,25 @@
+export type ModèleRéponse = {
+  data: {
+    suiviPar: string; // user qui édite le document
+    suiviParEmail: string; // email dgec var env
+    dreal: string; // région projet
+    refPotentiel: string; // identifiantProjet
+    nomRepresentantLegal: string;
+    nomCandidat: string;
+    status: string;
+    adresseCandidat: string;
+    email: string;
+    titrePeriode: string;
+    titreAppelOffre: string;
+    familles: 'yes' | '';
+    titreFamille: string;
+    nomProjet: string;
+    puissance: string;
+    codePostalProjet: string;
+    communeProjet: string;
+    unitePuissance: string;
+    dateDemande: string;
+    justificationDemande: string;
+    dateNotification: string;
+  };
+};
