+++ packages/applications/document-builder/src/modèleRéponseSignée/recours/modèleRéponseSignéeRecours.ts
@@ -1,0 +1,8 @@
+import { ModèleRéponse } from '../modèleRéponseSignée';
+
+export const modèleRéponseRecoursFileName = 'garanties-financières-modèle-mise-en-demeure.docx';
+
+export type ModèleRéponseRecours = ModèleRéponse & {
+  type: 'recours';
+  data: {};
+};
