+++ packages/applications/ssr/src/app/elimine/[identifiant]/recours/modele-reponse/route.ts
@@ -1,7 +1,7 @@
 import { mediator } from 'mediateur';
 import { notFound } from 'next/navigation';
 
-import { Abandon, CahierDesCharges } from '@potentiel-domain/laureat';
+import { Recours } from '@potentiel-domain/elimine';
 import { AppelOffre } from '@potentiel-domain/appel-offre';
 import { Candidature } from '@potentiel-domain/candidature';
 import { DateTime } from '@potentiel-domain/common';
@@ -42,14 +42,14 @@ export const GET = async (_: Request, { params: { identifiant } }: IdentifiantPa
       return notFound();
     }
 
-    const abandon = await mediator.send<Abandon.ConsulterAbandonQuery>({
-      type: 'Lauréat.Abandon.Query.ConsulterAbandon',
+    const recours = await mediator.send<Recours.ConsulterRecoursQuery>({
+      type: 'Éliminé.Recours.Query.ConsulterRecours',
       data: {
         identifiantProjetValue: identifiantProjet,
       },
     });
 
-    if (Option.isNone(abandon)) {
+    if (Option.isNone(recours)) {
       return notFound();
     }
 
@@ -62,49 +62,26 @@ export const GET = async (_: Request, { params: { identifiant } }: IdentifiantPa
       return notFound();
     }
 
-    const cahierDesChargesChoisi =
-      await mediator.send<CahierDesCharges.ConsulterCahierDesChargesChoisiQuery>({
-        type: 'Lauréat.CahierDesCharges.Query.ConsulterCahierDesChargesChoisi',
-        data: { identifiantProjet },
-      });
-
-    if (Option.isNone(cahierDesChargesChoisi)) {
-      return notFound();
-    }
-
-    const dispositionCDC = getCDCAbandonRefs({
-      appelOffres,
-      période: candidature.période,
-      cahierDesChargesChoisi,
-    });
-
     const content = await ModèleRéponseSignée.générerModèleRéponseAdapter({
-      type: 'abandon',
+      type: 'recours',
       data: {
-        aprèsConfirmation: abandon.demande.confirmation?.confirméLe ? true : false,
         adresseCandidat: candidature.candidat.adressePostale,
         codePostalProjet: candidature.localité.codePostal,
         communeProjet: candidature.localité.commune,
-        contenuParagrapheAbandon: dispositionCDC.dispositions,
-        dateConfirmation:
-          abandon.demande.confirmation?.confirméLe?.date.toLocaleDateString('fr-FR') || '',
-        dateDemande: abandon.demande.demandéLe.date.toLocaleDateString('fr-FR'),
-        dateDemandeConfirmation:
-          abandon.demande.confirmation?.demandéeLe.date.toLocaleDateString('fr-FR') || '',
+        dateDemande: recours.demande.demandéLe.date.toLocaleDateString('fr-FR'),
         dateNotification: DateTime.convertirEnValueType(
           candidature.dateDésignation,
         ).date.toLocaleDateString('fr-FR'),
         dreal: candidature.localité.région,
         email: '',
         familles: candidature.famille ? 'yes' : '',
-        justificationDemande: abandon.demande.raison,
+        justificationDemande: recours.demande.raison,
         nomCandidat: candidature.candidat.nom,
         nomProjet: candidature.nom,
         nomRepresentantLegal: candidature.candidat.représentantLégal,
         puissance: candidature.puissance.toString(),
-        referenceParagrapheAbandon: dispositionCDC.référenceParagraphe,
         refPotentiel: formatIdentifiantProjetForDocument(identifiantProjet),
-        status: abandon.statut.statut,
+        status: recours.statut.value,
         suiviPar: nomComplet || '',
         suiviParEmail: appelOffres.dossierSuiviPar,
         titreAppelOffre: appelOffres.title,
@@ -112,7 +89,6 @@ export const GET = async (_: Request, { params: { identifiant } }: IdentifiantPa
         titrePeriode:
           appelOffres.periodes.find((période) => période.id === candidature.période)?.title || '',
         unitePuissance: appelOffres.unitePuissance,
-        enCopies: getEnCopies(candidature.localité.région),
       },
     });
 
@@ -122,63 +98,3 @@ export const GET = async (_: Request, { params: { identifiant } }: IdentifiantPa
       },
     });
   });
-
-const getEnCopies = (region: string): Array<string> => {
-  if (!region) {
-    return ['DREAL concernée', 'CRE'];
-  }
-
-  const enCopie = [];
-
-  if (['Guadeloupe', 'Guyane', 'Martinique', 'Corse', 'La Réunion', 'Mayotte'].includes(region)) {
-    enCopie.push('EDF OA');
-  }
-
-  if (['Guadeloupe', 'Guyane', 'Martinique', 'Corse', 'La Réunion'].includes(region)) {
-    enCopie.push('EDF SEI');
-  }
-
-  if (['Mayotte'].includes(region)) {
-    enCopie.push('EDN');
-  }
-
-  return [...enCopie, `DREAL ${region}`, 'CRE'];
-};
-
-const getCDCAbandonRefs = ({
-  appelOffres,
-  période,
-  cahierDesChargesChoisi,
-}: {
-  appelOffres: AppelOffre.AppelOffreReadModel;
-  période: string;
-  cahierDesChargesChoisi: string;
-}) => {
-  const périodeDetails = appelOffres.periodes.find((periode) => periode.id === période);
-  const cdc = parseCahierDesChargesChoisi(cahierDesChargesChoisi);
-  const cahierDesChargesModifié = périodeDetails?.cahiersDesChargesModifiésDisponibles.find(
-    (c) => cdc.type === 'modifié' && c.paruLe === cdc.paruLe && c.alternatif === cdc.alternatif,
-  );
-
-  return {
-    référenceParagraphe: '!!!REFERENCE NON DISPONIBLE!!!',
-    dispositions: '!!!CONTENU NON DISPONIBLE!!!',
-    ...appelOffres.donnéesCourriersRéponse.texteEngagementRéalisationEtModalitésAbandon,
-    ...périodeDetails?.donnéesCourriersRéponse?.texteEngagementRéalisationEtModalitésAbandon,
-    ...(cahierDesChargesModifié &&
-      cahierDesChargesModifié.donnéesCourriersRéponse
-        ?.texteEngagementRéalisationEtModalitésAbandon),
-  };
-};
-
-const parseCahierDesChargesChoisi = (référence: string) => {
-  if (référence === 'initial') {
-    return { type: 'initial' };
-  }
-
-  return {
-    type: 'modifié',
-    paruLe: référence.replace('-alternatif', ''),
-    alternatif: référence.search('-alternatif') === -1 ? undefined : true,
-  };
-};
